Images added/deleted/changed + changes to main JS
Some images could be removed, since the projects were not actually done
or did not have anything to show for them. Also, added two projects
(Bean and Pepper's Ghost) and thus added two images.

Considering the code, I removed some of the javascript animations that
caused the automatic sliding of project and team slider-blocks. Also,
removed the part that caused the 'automatic scroll' to the bottom when
clicking a project or team member, since it was unnecessary and
unresponsive.

(CSS did not change except for one little  bracket that was out of line
with the rest. Sorry. I couldn't help myself.)