Fixes for the about, portfolio and contact pages
After Sunny changed the homepage 5 days ago, some problems occured on
the other pages. The website was built on pretty much every section
having the class 'block'. Since Sunny made changes to the CSS of this
class, it ruined the rest of the website.

Now every page has it's own block-class (teamblock, portfolioblock and
contactblock) so that we can change the styling on every page, without
changing the others. For now, the other 3 pages have the same CSS (the
old CSS) and only the homepage is different.